requirements.txt
attrs==22.2.0
Babel==2.11.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
exceptiongroup==1.1.0
Flask==2.2.2
flask-babel==3.0.1
Flask-Limiter==3.2.0
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.0
pluggy==1.0.0
Pygments==2.14.0
pytest==7.2.2
python-dotenv==0.21.1
pytz==2022.7.1
rich==13.3.1
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.13.0